Add area mapping and test data (#48)
* Change water_allocations.amount type to handle decimals

* Add allocation amount area mappings and test data

* Format code

+++ packages/IngestAPI/src/main/kotlin/nz/govt/eop/Messages.kt
@@ -1,11 +1,12 @@
 package nz.govt.eop.messages
 
+import java.math.BigDecimal
 import java.time.Instant
 import nz.govt.eop.ingestapi.WaterAllocation
 
 data class WaterAllocationMessage(
     val areaId: String,
-    val amount: Int,
+    val amount: BigDecimal,
     val ingestId: String,
     val receivedAt: Instant
 ) {
